Only add compactor if run in monolithic mode
This bases the decision whether we run in single binary mode on the
specified target. Making an assumption based on the lifecycler config
can lead to the wrong result. With the old behavior, compactor would
be started if running one replica of ruler with an `inmemory` ring even
when running in microservices mode, which is not desirable because
compactor is run separately and there should only be one compactor.

Signed-off-by: Reinhard Nägele <unguiculus@gmail.com>